chore: remove old session code and drop v2 naming
