--- src/conduit/shared/__init__.py
@@ -1,5 +1,0 @@
-"""Shared utilities for MCP client and server sessions."""
-
-from .session import BaseSession
-
-__all__ = ["BaseSession"]
